change structure of filesystem + o3, todo: makefile + .cc change

+++ cse5441_lab1/lab1.cpp
@@ -1,7 +1,6 @@
 #include <iostream>
 #include <cstdlib>
 #include <sys/time.h>
-#include <pthread.h>
 #include "lab1.h"
 using namespace std;
 
@@ -32,11 +31,10 @@ int main(int argc, char* argv[]) {
         gettimeofday(&t1, NULL);
         convergenceLoop();
         gettimeofday(&t2, NULL);
-        float diff = ((float)t2.tv_usec - (float)t1.tv_usec) / 1000000.0;
+        float diff = ((unsigned long long)t2.tv_usec - (unsigned long long)t1.tv_usec) / 1000000.0;
         //cout << "Loop " << loops << ": " << diff << " seconds!" << endl;
         runtime += diff;
     }
-
     cout << "Convergence Loops Complete. Freeing Memory..." << endl;
     freeBoxes();
     cout << "Number of Convergence Loops: " << loops << endl;
